refactor: get key value pairs to upload outside of upload_files fn
