+++ src/commands/kv/bucket/upload.rs
@@ -1,18 +1,12 @@
-use super::manifest::AssetManifest;
+use indicatif::ProgressBar;
 
-use std::collections::HashSet;
-use std::fs::metadata;
-use std::path::Path;
+use cloudflare::endpoints::workerskv::write_bulk::KeyValuePair;
+use cloudflare::framework::apiclient::ApiClient;
 
 use crate::commands::kv;
-use crate::commands::kv::bucket::directory_keys_values;
 use crate::settings::global_user::GlobalUser;
 use crate::settings::toml::Target;
 use crate::terminal::message;
-use cloudflare::endpoints::workerskv::write_bulk::KeyValuePair;
-use cloudflare::framework::apiclient::ApiClient;
-use failure::format_err;
-use indicatif::ProgressBar;
 
 // The consts below are halved from the API's true capacity to help avoid
 // hammering it with large requests.
@@ -23,25 +17,8 @@ pub fn upload_files(
     target: &Target,
     user: &GlobalUser,
     namespace_id: &str,
-    path: &Path,
-    ignore: &HashSet<String>,
-    verbose: bool,
-) -> Result<AssetManifest, failure::Error> {
-    let (mut pairs, asset_manifest): (Vec<KeyValuePair>, AssetManifest) = match &metadata(path) {
-        Ok(file_type) if file_type.is_dir() => {
-            let (pairs, asset_manifest) = directory_keys_values(target, path, verbose)?;
-            Ok((pairs, asset_manifest))
-        }
-
-        Ok(_file_type) => {
-            // any other file types (files, symlinks)
-            Err(format_err!("wrangler kv:bucket upload takes a directory"))
-        }
-        Err(e) => Err(format_err!("{}", e)),
-    }?;
-
-    pairs = filter_files(pairs, ignore);
-
+    mut pairs: Vec<KeyValuePair>,
+) -> Result<(), failure::Error> {
     if !pairs.is_empty() {
         let client = kv::api_client(user)?;
         // Iterate over all key-value pairs and create batches of uploads, each of which are
@@ -89,7 +66,7 @@ pub fn upload_files(
         }
     }
 
-    Ok(asset_manifest)
+    Ok(())
 }
 
 fn upload_batch(
@@ -106,84 +83,5 @@ fn upload_batch(
             Ok(())
         }
         Err(e) => failure::bail!("Failed to upload file batch. {}", kv::format_error(e)),
-    }
-}
-
-fn filter_files(pairs: Vec<KeyValuePair>, already_uploaded: &HashSet<String>) -> Vec<KeyValuePair> {
-    let mut filtered_pairs: Vec<KeyValuePair> = Vec::new();
-    for pair in pairs {
-        if !already_uploaded.contains(&pair.key) {
-            filtered_pairs.push(pair);
-        }
-    }
-    filtered_pairs
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::collections::HashSet;
-    use std::path::Path;
-
-    use crate::commands::kv::bucket::generate_path_and_key;
-    use cloudflare::endpoints::workerskv::write_bulk::KeyValuePair;
-
-    #[test]
-    fn it_can_filter_preexisting_files() {
-        let (_, key_a_old) =
-            generate_path_and_key(Path::new("/a"), Path::new("/"), Some("old".to_string()))
-                .unwrap();
-        let (_, key_b_old) =
-            generate_path_and_key(Path::new("/b"), Path::new("/"), Some("old".to_string()))
-                .unwrap();
-        // Generate new key (using hash of new value) for b when to simulate its value being updated.
-        let (_, key_b_new) =
-            generate_path_and_key(Path::new("/b"), Path::new("/"), Some("new".to_string()))
-                .unwrap();
-
-        // Old values found on remote
-        let mut exclude_keys = HashSet::new();
-        exclude_keys.insert(key_a_old.clone());
-        exclude_keys.insert(key_b_old);
-
-        // local files (with b updated) to upload
-        let pairs_to_upload = vec![
-            KeyValuePair {
-                key: key_a_old,
-                value: "old".to_string(), // This value remains unchanged
-                expiration_ttl: None,
-                expiration: None,
-                base64: None,
-            },
-            KeyValuePair {
-                key: key_b_new.clone(),
-                value: "new".to_string(), // Note this pair has a new value
-                expiration_ttl: None,
-                expiration: None,
-                base64: None,
-            },
-        ];
-
-        let expected = vec![KeyValuePair {
-            key: key_b_new,
-            value: "new".to_string(),
-            expiration_ttl: None,
-            expiration: None,
-            base64: None,
-        }];
-        let actual = filter_files(pairs_to_upload, &exclude_keys);
-        check_kv_pairs_equality(expected, actual);
-    }
-
-    fn check_kv_pairs_equality(expected: Vec<KeyValuePair>, actual: Vec<KeyValuePair>) {
-        assert!(expected.len() == actual.len());
-        for (idx, pair) in expected.into_iter().enumerate() {
-            // Ensure the expected key and value was returned in the filtered pair list
-            // Awkward field-by-field comparison below courtesy of not yet implementing
-            // PartialEq for KeyValuePair in cloudflare-rs :)
-            // TODO: (gabbi) Implement PartialEq for KeyValuePair in cloudflare-rs.
-            assert!(pair.key == actual[idx].key);
-            assert!(pair.value == actual[idx].value);
-        }
     }
 }
